pyproject.toml
[project]
name = "glacis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.3",
    "fastapi[standard]>=0.136.1",
    "google-genai>=1.74.0",
    "ollama>=0.6.2",
    "openai>=2.33.0",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "pydantic>=2.13.3",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
]
